Implement attempt history API (#14)
* Change attempt history table

* Implement attempt history

* Add user resolve field to quiz attempts

* Add quiz resolver to quiz  attempts

* Store more question info to quiz question attempt

* Refactor quiz attempt util

* Change pass score type to float

+++ server/src/quizzes/dto/create-quiz.input.ts
@@ -1,4 +1,4 @@
-import { InputType, Field, Int } from '@nestjs/graphql';
+import { InputType, Field, Float } from '@nestjs/graphql';
 import { IsInt, Length, Max, Min } from 'class-validator';
 import { CreateQuestionBaseInput } from 'src/questions/dto';
 
@@ -17,11 +17,11 @@ export class CreateQuizInput {
   @Field()
   level: string;
 
-  @Field(() => Int, { defaultValue: 70 })
+  @Field(() => Float, { defaultValue: 70 })
   @IsInt()
   @Min(50)
   @Max(100)
-  passScore: number;
+  passScore: number; // between 50% and 100%
 
   @Field(() => [CreateQuestionBaseInput], { nullable: true })
   questions?: CreateQuestionBaseInput[];
